changing up how the user edits the roomname
