+++ src/screens/message-board/MessageBoard.js
@@ -4,7 +4,7 @@ import OnlineUsers from '../../components/OnlineUsers'
 import Board from '../../components/Board'
 import MessageBoardForm from '../../components/MessageBoardForm'
 import DateTimeFluent from '../../components/DateTimeFluent'
-import Rooms from '../../components/Rooms'
+import Rooms from '../../components/Room/Rooms'
 import { useUser } from '../../context/user-context'
 import { useSocket } from '../../context/socket-context'
 import { useRoom } from '../../context/room-context'
